requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-test-without-migrations==0.6
djangorestframework==3.10.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.3.13
oauthlib==3.1.0
openapi-codec==1.3.2
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3

ipdb